The PID split task is the PID default task
