+++ Analysis/Tasks/PID/CMakeLists.txt
@@ -23,11 +23,6 @@ o2_add_dpl_workflow(pid-tof-qa-mc
                     PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase O2::AnalysisDataModel O2::AnalysisCore
                     COMPONENT_NAME Analysis)
 
-o2_add_dpl_workflow(pid-tof-split
-                    SOURCES pidTOF_split.cxx
-                    PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase O2::AnalysisDataModel O2::AnalysisCore
-                    COMPONENT_NAME Analysis)
-
 o2_add_dpl_workflow(pid-tof-tiny
                     SOURCES pidTOF_tiny.cxx
                     PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase O2::AnalysisDataModel O2::AnalysisCore
@@ -35,11 +30,6 @@ o2_add_dpl_workflow(pid-tof-tiny
 
 o2_add_dpl_workflow(pid-tpc
                     SOURCES pidTPC.cxx
-                    PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase O2::AnalysisDataModel O2::AnalysisCore
-                    COMPONENT_NAME Analysis)
-
-o2_add_dpl_workflow(pid-tpc-split
-                    SOURCES pidTPC_split.cxx
                     PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase O2::AnalysisDataModel O2::AnalysisCore
                     COMPONENT_NAME Analysis)
