+++ Analysis/Tasks/PID/pidTOF.cxx
@@ -8,43 +8,21 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
+///
+/// \file   pidTOF_split.cxx
+/// \author Nicolo' Jacazio
+/// \brief  Task to produce PID tables for TOF split for each particle.
+///         Only the tables for the mass hypotheses requested are filled, the others are sent empty.
+///
+
 // O2 includes
 #include "Framework/AnalysisTask.h"
-#include "Framework/AnalysisDataModel.h"
-#include "Framework/ASoAHelpers.h"
 #include "Framework/HistogramRegistry.h"
 #include "ReconstructionDataFormats/Track.h"
 #include <CCDB/BasicCCDBManager.h>
-#include "AnalysisDataModel/TrackSelectionTables.h"
 #include "AnalysisDataModel/PID/PIDResponse.h"
 #include "AnalysisDataModel/PID/PIDTOF.h"
-#include "AnalysisDataModel/PID/TOFReso.h"
 #include "AnalysisDataModel/TrackSelectionTables.h"
-
-// Table with the full information for all particles
-namespace o2::aod
-{
-
-DECLARE_SOA_TABLE(pidRespTOF, "AOD", "pidRespTOF",
-                  // Expected signals
-                  pidtof::TOFExpSignalDiffEl<pidtof::TOFNSigmaEl, pidtof::TOFExpSigmaEl>,
-                  pidtof::TOFExpSignalDiffMu<pidtof::TOFNSigmaMu, pidtof::TOFExpSigmaMu>,
-                  pidtof::TOFExpSignalDiffPi<pidtof::TOFNSigmaPi, pidtof::TOFExpSigmaPi>,
-                  pidtof::TOFExpSignalDiffKa<pidtof::TOFNSigmaKa, pidtof::TOFExpSigmaKa>,
-                  pidtof::TOFExpSignalDiffPr<pidtof::TOFNSigmaPr, pidtof::TOFExpSigmaPr>,
-                  pidtof::TOFExpSignalDiffDe<pidtof::TOFNSigmaDe, pidtof::TOFExpSigmaDe>,
-                  pidtof::TOFExpSignalDiffTr<pidtof::TOFNSigmaTr, pidtof::TOFExpSigmaTr>,
-                  pidtof::TOFExpSignalDiffHe<pidtof::TOFNSigmaHe, pidtof::TOFExpSigmaHe>,
-                  pidtof::TOFExpSignalDiffAl<pidtof::TOFNSigmaAl, pidtof::TOFExpSigmaAl>,
-                  // Expected sigma
-                  pidtof::TOFExpSigmaEl, pidtof::TOFExpSigmaMu, pidtof::TOFExpSigmaPi,
-                  pidtof::TOFExpSigmaKa, pidtof::TOFExpSigmaPr, pidtof::TOFExpSigmaDe,
-                  pidtof::TOFExpSigmaTr, pidtof::TOFExpSigmaHe, pidtof::TOFExpSigmaAl,
-                  // NSigma
-                  pidtof::TOFNSigmaEl, pidtof::TOFNSigmaMu, pidtof::TOFNSigmaPi,
-                  pidtof::TOFNSigmaKa, pidtof::TOFNSigmaPr, pidtof::TOFNSigmaDe,
-                  pidtof::TOFNSigmaTr, pidtof::TOFNSigmaHe, pidtof::TOFNSigmaAl);
-}
 
 using namespace o2;
 using namespace o2::framework;
@@ -54,24 +32,42 @@ using namespace o2::track;
 
 void customize(std::vector<o2::framework::ConfigParamSpec>& workflowOptions)
 {
-  std::vector<ConfigParamSpec> options{
-    {"use-fast", VariantType::Int, 0, {"Use fast implementation"}},
-    {"add-qa", VariantType::Int, 0, {"Produce TOF PID QA histograms"}}};
+  std::vector<ConfigParamSpec> options{{"add-qa", VariantType::Int, 0, {"Produce TOF PID QA histograms"}}};
   std::swap(workflowOptions, options);
 }
 
 #include "Framework/runDataProcessing.h"
 
-struct pidTOFTask {
+struct pidTOFTaskSplit {
   using Trks = soa::Join<aod::Tracks, aod::TracksExtra, aod::TracksCov>;
   using Coll = aod::Collisions;
-  Produces<aod::pidRespTOF> tablePID;
+  // Tables to produce
+  Produces<o2::aod::pidRespTOFEl> tablePIDEl;
+  Produces<o2::aod::pidRespTOFMu> tablePIDMu;
+  Produces<o2::aod::pidRespTOFPi> tablePIDPi;
+  Produces<o2::aod::pidRespTOFKa> tablePIDKa;
+  Produces<o2::aod::pidRespTOFPr> tablePIDPr;
+  Produces<o2::aod::pidRespTOFDe> tablePIDDe;
+  Produces<o2::aod::pidRespTOFTr> tablePIDTr;
+  Produces<o2::aod::pidRespTOFHe> tablePIDHe;
+  Produces<o2::aod::pidRespTOFAl> tablePIDAl;
+  // Detector response and input parameters
   DetectorResponse response;
   Service<o2::ccdb::BasicCCDBManager> ccdb;
   Configurable<std::string> paramfile{"param-file", "", "Path to the parametrization object, if emtpy the parametrization is not taken from file"};
   Configurable<std::string> sigmaname{"param-sigma", "TOFReso", "Name of the parametrization for the expected sigma, used in both file and CCDB mode"};
   Configurable<std::string> url{"ccdb-url", "http://ccdb-test.cern.ch:8080", "url of the ccdb repository"};
   Configurable<long> timestamp{"ccdb-timestamp", -1, "timestamp of the object"};
+  // Configuration flags to include and exclude particle hypotheses
+  Configurable<int> pidEl{"pid-el", 0, {"Produce PID information for the Electron mass hypothesis"}};
+  Configurable<int> pidMu{"pid-mu", 0, {"Produce PID information for the Muon mass hypothesis"}};
+  Configurable<int> pidPi{"pid-pi", 0, {"Produce PID information for the Pion mass hypothesis"}};
+  Configurable<int> pidKa{"pid-ka", 0, {"Produce PID information for the Kaon mass hypothesis"}};
+  Configurable<int> pidPr{"pid-pr", 0, {"Produce PID information for the Proton mass hypothesis"}};
+  Configurable<int> pidDe{"pid-de", 0, {"Produce PID information for the Deuterons mass hypothesis"}};
+  Configurable<int> pidTr{"pid-tr", 0, {"Produce PID information for the Triton mass hypothesis"}};
+  Configurable<int> pidHe{"pid-he", 0, {"Produce PID information for the Helium3 mass hypothesis"}};
+  Configurable<int> pidAl{"pid-al", 0, {"Produce PID information for the Alpha mass hypothesis"}};
 
   void init(o2::framework::InitContext&)
   {
@@ -86,7 +82,6 @@ struct pidTOFTask {
     response.SetParameters(DetectorResponse::kSigma, p);
     const std::string fname = paramfile.value;
     if (!fname.empty()) { // Loading the parametrization from file
-      LOG(INFO) << "Loading parametrization from file" << fname << ", using param: " << sigmaname;
       response.LoadParamFromFile(fname.data(), sigmaname.value, DetectorResponse::kSigma);
     } else { // Loading it from CCDB
       const std::string path = "Analysis/PID/TOF";
@@ -108,97 +103,26 @@ struct pidTOFTask {
     constexpr auto responseHe = ResponseImplementation<PID::Helium3>();
     constexpr auto responseAl = ResponseImplementation<PID::Alpha>();
 
-    tablePID.reserve(tracks.size());
-    for (auto const& trk : tracks) {
-      tablePID(responseEl.GetExpectedSigma(response, trk.collision(), trk),
-               responseMu.GetExpectedSigma(response, trk.collision(), trk),
-               responsePi.GetExpectedSigma(response, trk.collision(), trk),
-               responseKa.GetExpectedSigma(response, trk.collision(), trk),
-               responsePr.GetExpectedSigma(response, trk.collision(), trk),
-               responseDe.GetExpectedSigma(response, trk.collision(), trk),
-               responseTr.GetExpectedSigma(response, trk.collision(), trk),
-               responseHe.GetExpectedSigma(response, trk.collision(), trk),
-               responseAl.GetExpectedSigma(response, trk.collision(), trk),
-               responseEl.GetSeparation(response, trk.collision(), trk),
-               responseMu.GetSeparation(response, trk.collision(), trk),
-               responsePi.GetSeparation(response, trk.collision(), trk),
-               responseKa.GetSeparation(response, trk.collision(), trk),
-               responsePr.GetSeparation(response, trk.collision(), trk),
-               responseDe.GetSeparation(response, trk.collision(), trk),
-               responseTr.GetSeparation(response, trk.collision(), trk),
-               responseHe.GetSeparation(response, trk.collision(), trk),
-               responseAl.GetSeparation(response, trk.collision(), trk));
-    }
-  }
-};
-
-struct pidTOFTaskFast {
-  using Trks = soa::Join<aod::Tracks, aod::TracksExtra, aod::TracksCov>;
-  using Coll = aod::Collisions;
-  Produces<aod::pidRespTOF> tablePID;
-  DetectorResponse response;
-  Parameters resoParameters;
-  Service<o2::ccdb::BasicCCDBManager> ccdb;
-  Configurable<std::string> paramfile{"param-file", "", "Path to the parametrization object, if emtpy the parametrization is not taken from file"};
-  Configurable<std::string> sigmaname{"param-sigma", "TOFReso", "Name of the parametrization for the expected sigma, used in both file and CCDB mode"};
-  Configurable<std::string> url{"ccdb-url", "http://ccdb-test.cern.ch:8080", "url of the ccdb repository"};
-  Configurable<long> timestamp{"ccdb-timestamp", -1, "timestamp of the object"};
-
-  void init(o2::framework::InitContext&)
-  {
-    ccdb->setURL(url.value);
-    ccdb->setTimestamp(timestamp.value);
-    ccdb->setCaching(true);
-    ccdb->setLocalObjectValidityChecking();
-    // Not later than now objects
-    ccdb->setCreatedNotAfter(std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count());
-    //
-    const std::vector<float> p = {0.008, 0.008, 0.002, 40.0};
-    response.SetParameters(DetectorResponse::kSigma, p);
-    const std::string fname = paramfile.value;
-    if (!fname.empty()) { // Loading the parametrization from file
-      LOG(INFO) << "Loading parametrization from file" << fname << ", using param: " << sigmaname;
-      response.LoadParamFromFile(fname.data(), sigmaname.value, DetectorResponse::kSigma);
-    } else { // Loading it from CCDB
-      const std::string path = "Analysis/PID/TOF";
-      response.LoadParam(DetectorResponse::kSigma, ccdb->getForTimeStamp<Parametrization>(path + "/" + sigmaname.value, timestamp.value));
-    }
-    resoParameters = response.GetParam(DetectorResponse::kSigma)->GetParameters();
-  }
-
-  template <o2::track::PID::ID id>
-  float sigma(Trks::iterator track)
-  {
-    return o2::pid::tof::TOFResoParamTrack<id>(track.collision(), track, resoParameters);
-  }
-  template <o2::track::PID::ID id>
-  float nsigma(Trks::iterator track)
-  {
-    return (track.tofSignal() - track.collision().collisionTime() * 1000.f - tof::ExpTimes<Coll::iterator, Trks::iterator, id>::GetExpectedSignal(track.collision(), track)) / sigma<id>(track);
-  }
-  void process(Coll const& collisions, Trks const& tracks)
-  {
-    tablePID.reserve(tracks.size());
-    for (auto const& trk : tracks) {
-      tablePID(sigma<PID::Electron>(trk),
-               sigma<PID::Muon>(trk),
-               sigma<PID::Pion>(trk),
-               sigma<PID::Kaon>(trk),
-               sigma<PID::Proton>(trk),
-               sigma<PID::Deuteron>(trk),
-               sigma<PID::Triton>(trk),
-               sigma<PID::Helium3>(trk),
-               sigma<PID::Alpha>(trk),
-               nsigma<PID::Electron>(trk),
-               nsigma<PID::Muon>(trk),
-               nsigma<PID::Pion>(trk),
-               nsigma<PID::Kaon>(trk),
-               nsigma<PID::Proton>(trk),
-               nsigma<PID::Deuteron>(trk),
-               nsigma<PID::Triton>(trk),
-               nsigma<PID::Helium3>(trk),
-               nsigma<PID::Alpha>(trk));
-    }
+    // Check and fill enabled tables
+    auto makeTable = [&tracks](const Configurable<int>& flag, auto& table, const DetectorResponse& response, const auto& responsePID) {
+      if (flag.value) {
+        // Prepare memory for enabled tables
+        table.reserve(tracks.size());
+        for (auto const& trk : tracks) { // Loop on Tracks
+          table(responsePID.GetExpectedSigma(response, trk.collision(), trk),
+                responsePID.GetSeparation(response, trk.collision(), trk));
+        }
+      }
+    };
+    makeTable(pidEl, tablePIDEl, response, responseEl);
+    makeTable(pidMu, tablePIDMu, response, responseMu);
+    makeTable(pidPi, tablePIDPi, response, responsePi);
+    makeTable(pidKa, tablePIDKa, response, responseKa);
+    makeTable(pidPr, tablePIDPr, response, responsePr);
+    makeTable(pidDe, tablePIDDe, response, responseDe);
+    makeTable(pidTr, tablePIDTr, response, responseTr);
+    makeTable(pidHe, tablePIDHe, response, responseHe);
+    makeTable(pidAl, tablePIDAl, response, responseAl);
   }
 };
 
@@ -298,7 +222,11 @@ struct pidTOFTaskQA {
     histos.fill(HIST(hnsigma[i]), t.p(), nsigma);
   }
 
-  void process(aod::Collision const& collision, soa::Join<aod::Tracks, aod::TracksExtra, aod::TracksCov, aod::pidRespTOF, aod::TrackSelection> const& tracks)
+  void process(aod::Collision const& collision, soa::Join<aod::Tracks, aod::TracksExtra, aod::TracksCov,
+                                                          aod::pidRespTOFEl, aod::pidRespTOFMu, aod::pidRespTOFPi,
+                                                          aod::pidRespTOFKa, aod::pidRespTOFPr, aod::pidRespTOFDe,
+                                                          aod::pidRespTOFTr, aod::pidRespTOFHe, aod::pidRespTOFAl,
+                                                          aod::TrackSelection> const& tracks)
   {
     const float collisionTime_ps = collision.collisionTime() * 1000.f;
     histos.fill(HIST("event/vertexz"), collision.posZ());
@@ -337,12 +265,7 @@ struct pidTOFTaskQA {
 
 WorkflowSpec defineDataProcessing(ConfigContext const& cfgc)
 {
-  auto workflow = WorkflowSpec{};
-  if (cfgc.options().get<int>("use-fast")) {
-    workflow.push_back(adaptAnalysisTask<pidTOFTaskFast>(cfgc, TaskName{"pidTOF-task"}));
-  } else {
-    workflow.push_back(adaptAnalysisTask<pidTOFTask>(cfgc, TaskName{"pidTOF-task"}));
-  }
+  auto workflow = WorkflowSpec{adaptAnalysisTask<pidTOFTaskSplit>(cfgc, TaskName{"pidTOF-split-task"})};
   if (cfgc.options().get<int>("add-qa")) {
     workflow.push_back(adaptAnalysisTask<pidTOFTaskQA>(cfgc, TaskName{"pidTOFQA-task"}));
   }
